Move load balancing code in its own file.
